# FlatGeoBuf and GPKG: improve DateTime parsing perf + GPKG improve GetNextArrowArray() perf