… ogr/ogrsf_frmts/gpkg/ogrgeopackagelayer.cpp
… ogr/ogrutils.cpp
… swig/include/python/ogr_python.i
… swig/python/osgeo/ogr.py