With those improvements, the timing of [bench_ogr_batch](https://github.com/OSGeo/gdal/blob/3f5cd1cc33685a627ee786bccaa0bf312d394ad9/apps/bench_ogr_batch.cpp), which uses ArrowArray batch reading, on GPKG and FlatGeoBuf files converted from https://storage.googleapis.com/open-geodata/linz-examples/nz-building-outlines.parquet goes from:

- FlatGeoBuf: 4.5 s -> 3.2 s
- GPKG: 4.5 s -> 2.2 s (single-threaded) -> 1.5 s (main thread + worker thread)

compared to 1.4 s on the GeoParquet file (whose reading uses threading)